+++ CMakeLists.txt
@@ -4,18 +4,18 @@ cmake_minimum_required(VERSION 3.1 FATAL_ERROR)
 ## These variables are passed to oatpp-module-install.cmake script
 ## use these variables to configure module installation
 
-set(OATPP_THIS_MODULE_NAME oatpp-my-module) ## name of the module (also name of folders in installation dirs)
+set(OATPP_THIS_MODULE_NAME oatpp-wolfssl) ## name of the module (also name of folders in installation dirs)
 set(OATPP_THIS_MODULE_VERSION "1.3.0") ## version of the module (also sufix of folders in installation dirs)
-set(OATPP_THIS_MODULE_LIBRARIES oatpp-my-module) ## list of libraries to find when find_package is called
-set(OATPP_THIS_MODULE_TARGETS oatpp-my-module) ## list of targets to install
-set(OATPP_THIS_MODULE_DIRECTORIES oatpp-my-module) ## list of directories to install
+set(OATPP_THIS_MODULE_LIBRARIES oatpp-wolfssl) ## list of libraries to find when find_package is called
+set(OATPP_THIS_MODULE_TARGETS oatpp-wolfssl) ## list of targets to install
+set(OATPP_THIS_MODULE_DIRECTORIES oatpp-wolfssl) ## list of directories to install
 
 ###################################################################################################
 
 project(${OATPP_THIS_MODULE_NAME}
         VERSION ${OATPP_THIS_MODULE_VERSION}
         LANGUAGES CXX
-        ## HOMEPAGE_URL "https://github.com/oatpp/oatpp-my-module"
+        ## HOMEPAGE_URL "https://github.com/oatpp/oatpp-wolfssl"
         ## DESCRIPTION "Something about my-module"
 )
