+++ README.md
@@ -1,4 +1,4 @@
-# wolfssl
-TLS adaptor for wolfSSL
+# oatpp-wolfssl
+Oatpp TLS adaptor for wolfSSL.
 
 # THIS MODULE IS NOT READY TO USE
