+++ src/oatpp-wolfssl/MyClass.hpp
